third_party/blink/web_tests/svg/text/lengthAdjust-text-metrics.html

<!DOCTYPE HTML PUBLIC "-//IETF//DTD HTML//EN">
<html>
<head>
<!-- Test that assures that SVGTextContentElement API takes lengthAdjust & co into account. -->
<title>Complete test of the SVGTextContentElement API</title>
<script src="../../resources/ahem.js"></script>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
</head>
<body>
<script>
// Prepare testcase.
setup(function() {
    var svgNS = "http://www.w3.org/2000/svg";

    window.svgRoot = document.createElementNS(svgNS, "svg");
    document.documentElement.appendChild(svgRoot);

    window.svgText = document.createElementNS(svgNS, "text");
    svgText.style.fontFamily = "Ahem";
    svgText.style.fontSize = "20px";
    svgText.setAttribute("x", "10");
    svgText.setAttribute("y", "20");
    svgText.setAttribute("lengthAdjust", "spacingAndGlyphs");
    svgText.setAttribute("textLength", "200");
    svgText.appendChild(document.createTextNode("Test"));
    svgRoot.appendChild(svgText);

    // Compute scale matrix for spacingAndGlyphs.
    var startPosition = svgText.getStartPositionOfChar(0);
    window.scale = svgText.textLength.baseVal.value / svgText.getComputedTextLength();
    window.scaleMatrix = svgRoot.createSVGMatrix()
                                .translate(startPosition.x, startPosition.y)
                                .scaleNonUniform(scale, 1)
                                .translate(-startPosition.x, -startPosition.y);
    window.inverseScaleMatrix = scaleMatrix.inverse();
});

function transformPoint(point, matrix) {
    return point.matrixTransform(matrix);
}

function transformRect(rect, matrix) {
    var topLeft = svgRoot.createSVGPoint();
    topLeft.x = rect.x;
    topLeft.y = rect.y;
    topLeft = transformPoint(topLeft, matrix);

    var bottomRight = svgRoot.createSVGPoint();
    bottomRight.x = rect.x + rect.width;
    bottomRight.y = rect.y + rect.height;
    bottomRight = transformPoint(bottomRight, matrix);

    var newRect = svgRoot.createSVGRect();
    newRect.x = topLeft.x;
    newRect.y = topLeft.y;
    newRect.width = bottomRight.x - topLeft.x;
    newRect.height = bottomRight.y - topLeft.y;
    return newRect;
}

// Assertion helpers.
function assert_point_equals(actual, expected, description) {
    assert_approx_equals(actual.x, expected.x, 0.1, description + " (x)");
    assert_approx_equals(actual.y, expected.y, 0.1, description + " (y)");
}

function assert_rect_equals(actual, expected, description) {
    assert_approx_equals(actual.x, expected.x, 0.1, description + " (x)");
    assert_approx_equals(actual.y, expected.y, 0.1, description + " (y)");
    assert_approx_equals(actual.width, expected.width, 0.1, description + " (width)");
    assert_approx_equals(actual.height, expected.height, 0.1, description + " (height)");
}

// === Subtests ===

test(function() {
    assert_approx_equals(svgText.textLength.baseVal.value, 200, 0.1, "textLength");
    assert_equals(svgText.lengthAdjust.baseVal, SVGTextContentElement.LENGTHADJUST_SPACINGANDGLYPHS, "lengthAdjust");
}, "SVG DOM textLength and lengthAdjust");

test(function() {
    assert_equals(svgText.getNumberOfChars(), 4);
}, "getNumberOfChars");

test(function() {
    assert_approx_equals(svgText.getComputedTextLength() * scale, 200.0, 0.1, "getComputedTextLength");
}, "getComputedTextLength");

test(function() {
    assert_approx_equals(svgText.getSubStringLength(0, 1) * scale, 50, 0.1, "getSubStringLength(0, 1)");
    assert_approx_equals(svgText.getSubStringLength(0, 2) * scale, 100, 0.1, "getSubStringLength(0, 2)");
    assert_approx_equals(svgText.getSubStringLength(0, 3) * scale, 150, 0.1, "getSubStringLength(0, 3)");
    assert_approx_equals(svgText.getSubStringLength(0, 4) * scale, 200, 0.1, "getSubStringLength(0, 4)");
    assert_approx_equals(svgText.getSubStringLength(1, 1) * scale, 50, 0.1, "getSubStringLength(1, 1)");
    assert_approx_equals(svgText.getSubStringLength(1, 2) * scale, 100, 0.1, "getSubStringLength(1, 2)");
    assert_approx_equals(svgText.getSubStringLength(1, 3) * scale, 150, 0.1, "getSubStringLength(1, 3)");
    assert_approx_equals(svgText.getSubStringLength(2, 1) * scale, 50, 0.1, "getSubStringLength(2, 1)");
    assert_approx_equals(svgText.getSubStringLength(2, 2) * scale, 100, 0.1, "getSubStringLength(2, 2)");
    assert_approx_equals(svgText.getSubStringLength(3, 1) * scale, 50, 0.1, "getSubStringLength(3, 1)");
}, "getSubStringLength");

test(function() {
    var expected = [
        {x:10, y:20}, {x:60, y:20}, {x:110, y:20}, {x:160, y:20}
    ];
    for (var i = 0; i < 4; ++i) {
        var pt = svgText.getStartPositionOfChar(i);
        pt = pt.matrixTransform(scaleMatrix);
        assert_point_equals(pt, expected[i], "getStartPositionOfChar(" + i + ")");
    }
}, "getStartPositionOfChar");

test(function() {
    var expected = [
        {x:60, y:20}, {x:110, y:20}, {x:160, y:20}, {x:210, y:20}
    ];
    for (var i = 0; i < 4; ++i) {
        var pt = svgText.getEndPositionOfChar(i);
        pt = pt.matrixTransform(scaleMatrix);
        assert_point_equals(pt, expected[i], "getEndPositionOfChar(" + i + ")");
    }
}, "getEndPositionOfChar");

test(function() {
    var expected = [
        {x:10, y:4, width:50, height:20},
        {x:60, y:4, width:50, height:20},
        {x:110, y:4, width:50, height:20},
        {x:160, y:4, width:50, height:20}
    ];
    for (var i = 0; i < 4; ++i) {
        var rect = svgText.getExtentOfChar(i);
        rect = transformRect(rect, scaleMatrix);
        assert_rect_equals(rect, expected[i], "getExtentOfChar(" + i + ")");
    }
}, "getExtentOfChar");

test(function() {
    for (var i = 0; i < 4; ++i) {
        assert_approx_equals(svgText.getRotationOfChar(i), 0.0, 0.1, "getRotationOfChar(" + i + ")");
    }
}, "getRotationOfChar");

test(function() {
    // Points to test, with expected char index.
    var cases = [
        {x:0.0, expected: -1},
        {x:9.9, expected: -1},
        {x:10.1, expected: 0},
        {x:49.4, expected: 0},
        {x:71.6, expected: 1},
        {x:99.8, expected: 1},
        {x:128.0, expected: 2},
        {x:149.1, expected: 2},
        {x:179.3, expected: 3},
        {x:209.9, expected: 3},
        {x:210.1, expected: -1},
        {x:250.0, expected: -1},
    ];
    for (var i = 0; i < cases.length; ++i) {
        var point = svgRoot.createSVGPoint();
        point.x = cases[i].x;
        point.y = 10.0;
        point = point.matrixTransform(inverseScaleMatrix);
        assert_equals(svgText.getCharNumAtPosition(point), cases[i].expected, "getCharNumAtPosition(" + cases[i].x + ")");
    }
}, "getCharNumAtPosition");

// Cleanup.
document.documentElement.removeChild(svgRoot);
</script>
</body>
</html>
